Some updates
